backend/pyproject.toml
[project]
name = "wordle-benchmark"
version = "0.1.0"
description = "Testing LLM performance with Wordle"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]

[tool.pytest.ini_options]
markers = [
    "integration: marks tests as integration tests",
    "api_calls: mark tests that make real API calls (expensive)"
]
